+++ lib/src/codecs/codecs.dart
@@ -27,3 +27,4 @@ export 'datetime/xsd_datetime_codec.dart';
 export 'any_uri/any_uri_codec.dart';
 export 'decimal/decimal_codec.dart';
 export 'double/double_codec.dart';
+export 'float/float_codec.dart';