+++ CHANGELOG.md
@@ -30,3 +30,4 @@
 - Implement `xsd:anyUri` codec
 - Implement `xsd:decimal` codec
 - Implement `xsd:double` codec
+- Implement `xsd:float` codec